Termino UserController, creo UserService y UserServiceImpl. y creo RoomRepository.

+++ src/main/java/com/escuadronSuicida/backend/models/Room.java
@@ -17,7 +17,7 @@ public class Room {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private Double capacity;
+    private Integer capacity;
     private Boolean hasSockets;
 
     // @OneToMany con Keynotes
